collaborative_cookbook_backend/recipes/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model
from .models import Ingredient, Recipe

# CustomUser Serializer
class CustomUserSerializer(serializers.ModelSerializer):
    class Meta:
        model = get_user_model()
        fields = ('id', 'username', 'email', 'first_name', 'last_name')

class IngredientSerializer(serializers.ModelSerializer):
    class Meta:
        model = Ingredient
        fields = ('id', 'name',)

class RecipeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Recipe
        fields = ('id', 'name', 'description', 'instructions', 'image', 'ingredients', 'created_by')
