collaborative_cookbook_backend/requirements.txt
aniso8601==9.0.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
decorator==5.1.1
Django==4.2.4
django-cors-headers==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
executing==1.2.0
graphene==3.3
graphene-django==3.1.5
graphene-file-upload==1.3.0
graphql-core==3.2.3
graphql-relay==3.2.0
ipython==8.13.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
promise==2.3
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-decouple==3.8
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
text-unidecode==1.3
traitlets==5.9.0
wcwidth==0.2.6
